+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Check for proper width/height/depth when decoding alpha with dav1d, matching libaom's impl
 
 ## [0.3.6] - 2019-07-25
 ### Added